Anime class and engine creation moved to models.py
